here is your push go for it !

+++ Source/GC_UE4CPP/MyGameStateBase.h
@@ -18,9 +18,9 @@ public:
 	virtual ~AMyGameStateBase() = default;
 
 	UPROPERTY(EditAnywhere,BlueprintReadWrite)
-		int32 NumberOfSteaks;
+		float NumberOfSteaks;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-		int32 NumberOfAI;
+		float NumberOfAI;
 
 };
